+++ packages/plugins/src/plugins/meteora/dlmm/dlmmVaultsJob.ts
@@ -2,20 +2,20 @@ import { NetworkId } from '@sonarwatch/portfolio-core';
 import { Cache } from '../../../Cache';
 import { Job, JobExecutor } from '../../../Job';
 import { getClientSolana } from '../../../utils/clients';
-import { getParsedProgramAccounts } from '../../../utils/solana';
-import { dataSizeFilter } from '../../../utils/solana/filters';
 import { dlmmVaultProgramId, dlmmVaultsKey, platformId } from '../constants';
 import { dlmmVaultStruct } from './structs';
 import { CachedDlmmVaults } from '../types';
+import { ParsedGpa } from '../../../utils/solana/beets/ParsedGpa';
 
 const executor: JobExecutor = async (cache: Cache) => {
   const client = getClientSolana();
-  const accounts = await getParsedProgramAccounts(
+  const accounts = await ParsedGpa.build(
     client,
     dlmmVaultStruct,
-    dlmmVaultProgramId,
-    dataSizeFilter(dlmmVaultStruct.byteSize)
-  );
+    dlmmVaultProgramId
+  )
+    .addFilter('accountDiscriminator', [211, 8, 232, 43, 2, 152, 117, 119])
+    .run();
   const vaults: CachedDlmmVaults = {};
   accounts.forEach((acc) => {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any